// src/types/fcm/md/partner/BcncRegist/BcncRegist.types.ts
export interface BcncShipResponse {
  /** 배송지ID */
  shipId?: string;
  /** 배송지거래처코드 */
  shipCustno?: string;
  /** 배송지명 */
  shipName?: string;
  /** 주요배송지여부 */
  primaryChk?: string;
  /** 배송지주소 */
  shipAddr?: string;
  /** 영업사원 */
  salesMan?: string;
  /** 영업사원명 */
  salesName?: string;
  /** 사용여부 */
  useYn?: string;
  /** 채널1 */
  channel?: string;
  /** 채널2 */
  channel2?: string;
  /** 채널3 */
  channel3?: string;
  /** Territory1 */
  category1?: string;
  /** Territory2 */
  category2?: string;
  /** Territory3 */
  category3?: string;
  /** Territory4 */
  category4?: string;
  /** 국가 */
  country?: string;
  /** 사업장 */
  orgId?: string;
  /** 국가명 */
  nationName?: string;
  /** 화폐단위 */
  currency?: string;
  /** 담당자 */
  chargeMan?: string;
  /** 전화번호 */
  phoneNum?: string;
  /** 담당자번호 */
  chargeNumber?: string;
  /** 팩스번호 */
  siteFaxNo?: string;
  /** 계약일 */
  contractFrom?: string;
  /** 계약만료일 */
  contractTo?: string;
  /** 비고 */
  attribute1?: string;
  /** 메모 */
  attribute2?: string;
}